Lot's of changes, check description for more info.
- Complete command system overhaul.
-- Commands now each have their own class.
-- Commands now remind you what the command looks like on wrong arguments.
-- TabCompleter for boolean added.
-- /lpr, /lpr help, /lpr reload and /lpr reset commands added.

- Times now get saved as list of integers.
- New eligible field added to Reward instead of -1 in time.
- timeTillNextReward is initialized with new array of timetillnextreward of configreward instead of reference.
- Check on data reading if timetillnextreward is -1, if so change eligible to false.
- Removed singleton pattern in main class.
- getRewardsFromConfig now uses build in YAML features instead of making own ones.
- playtime and afktime are now integers.
- countdown check is more performant.
- Changed color of error messages.

+++ src/com/backtobedrock/LitePlaytimeRewards/helperClasses/ConfigReward.java
@@ -8,7 +8,7 @@ public class ConfigReward {
     private final String displayName;
     private final Material displayItem;
     private final List<String> displayDescription;
-    private final List<Long> playtimeNeeded;
+    private final List<Integer> playtimeNeeded;
     private final boolean countAfkTime;
     private final int slotsNeeded;
     private final boolean loop;
@@ -19,7 +19,7 @@ public class ConfigReward {
     private final String broadcastNotification;
     private final List<String> commands;
 
-    public ConfigReward(String displayName, Material displayItem, List<String> displayDescription, List<Long> playtimeNeeded, boolean countAfkTime, int slotsNeeded, boolean loop, List<String> disabledWorlds, boolean UsePermission, String notificationType, String notification, String broadcastNotification, List<String> commands) {
+    public ConfigReward(String displayName, Material displayItem, List<String> displayDescription, List<Integer> playtimeNeeded, boolean countAfkTime, int slotsNeeded, boolean loop, List<String> disabledWorlds, boolean UsePermission, String notificationType, String notification, String broadcastNotification, List<String> commands) {
         this.displayName = displayName;
         this.displayItem = displayItem;
         this.displayDescription = displayDescription;
@@ -43,7 +43,7 @@ public class ConfigReward {
         return displayName;
     }
 
-    public List<Long> getPlaytimeNeeded() {
+    public List<Integer> getPlaytimeNeeded() {
         return playtimeNeeded;
     }
